--- modules/create_teacher.rb
@@ -1,17 +1,0 @@
-module CreateTeacher
-  def create_teacher
-    puts "\n"
-    puts 'Please provide these information :'
-    print 'Age : '
-    age = gets.chomp
-    print 'Name : '
-    name = gets.chomp
-    print 'specialization : '
-    specialization = gets.chomp
-
-    new_teacher = Teacher.new(age, specialization, name)
-    @people_list << new_teacher
-    puts 'Teacher created successfully'
-    puts "\n"
-  end
-end
